refactor : Repository에 SQL문을 DAO로 리팩토링
